+++ include/boost/beast/http/write.hpp
@@ -10,6 +10,7 @@
 #ifndef BOOST_BEAST_HTTP_WRITE_HPP
 #define BOOST_BEAST_HTTP_WRITE_HPP
 
+#include <boost/beast/http/basic_file_body_fwd.hpp>
 #include <boost/beast/core/detail/config.hpp>
 #include <boost/beast/http/message.hpp>
 #include <boost/beast/http/serializer.hpp>